requirements.txt
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.7.0
durable-rules==2.0.28
fonttools==4.55.1
Jinja2==3.1.4
MarkupSafe==3.0.2
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
tinycss2==1.4.0
tinyhtml5==2.0.0
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3.post1
